bank_statement_chunker/requirements.txt
pandas>=1.5.0
pdfplumber>=0.7.0
openpyxl>=3.0.0
xlrd>=2.0.0
python-multipart>=0.0.6
fastapi>=0.68.0
uvicorn>=0.15.0
python-toon